refactor(hooks): consolidate useAppActions hook
