+++ src/hooks/actions/useLibraryActions.ts
@@ -5,7 +5,7 @@ import { useCallback } from 'react';
 import { fileOpen, fileSave } from 'browser-fs-access';
 import type { AnyPath, StyleClipboardData, MaterialData, LibraryData, RectangleData, ImageData, PolygonData, TextData, Point, GroupData, ArcData } from '../../types';
 import { getPathsBoundingBox, movePath } from '../../lib/drawing';
-import type { AppActionsProps } from '../useAppActions';
+import type { AppActionsProps } from './useAppActions';
 
 /**
  * 封装样式库和素材库相关操作的 Hook。
