+++ src/hooks/actions/useObjectActions.ts
@@ -4,7 +4,7 @@
 import { useCallback, useRef } from 'react';
 import { rectangleToVectorPath, ellipseToVectorPath, lineToVectorPath, brushToVectorPath, polygonToVectorPath, arcToVectorPath, flipPath, getPathsBoundingBox, alignPaths, distributePaths, performBooleanOperation, scalePath, movePath } from '../../lib/drawing';
 import type { AnyPath, RectangleData, EllipseData, VectorPathData, BrushPathData, PolygonData, ArcData, GroupData, Alignment, DistributeMode, ImageData, TextData, TraceOptions } from '../../types';
-import type { AppActionsProps } from '../useAppActions';
+import type { AppActionsProps } from './useAppActions';
 import { importSvg } from '../../lib/import';
 import { removeBackground, adjustHsv, type HsvAdjustment } from '../../lib/image';
